feat: add exported date constraints to dateUtilities for external reference by consumers and allow for barrel imports
