+++ scripts/runtime/dateUtilities/dateConstraints.mjs
@@ -1,0 +1,11 @@
+// filepath: dateConstraints.mjs
+const DATE_UTIL_CONSTRAINTS = {
+  maxDay: 31,
+  maxMonth: 12,
+  maxYear: 2400,
+  minDay: 1,
+  minMonth: 1,
+  minYear: 1900,
+};
+
+export { DATE_UTIL_CONSTRAINTS };
